Update question5.py
Improved Version of Code:
Improved Version of Code:

Here I used : 
Format of the code: 

value_when_true if condition else value_when_false

so in place of:
value_when_true=print("True")
if condition=lst_1[0] == lst_1[len(lst_1)-1]
value_when_false=print("False")

so, instead of 4 lines, in one line your code is reduced. So, now your code is of 2 lines.

This is not an advanced version, it is one way to write your code and can be used where the one-line statement is there for if & else condition.

Hope you will like it.